FEAT: Se agregan modificaciones al backend

+++ server/models/materiaPrima.model.js
@@ -15,7 +15,7 @@ module.exports = (sequelize, DataType) => {
       type: DataType.STRING,
     },
     espesor: {
-      type: DataType.INTEGER,
+      type: DataType.DOUBLE,
     },
   },
     {
